Update to Kotlin 1.9.20-Beta2
Since expect/actual classes are in Beta, their use has been removed as
much as reasonably possible.

The removals in the public API are probably because of changes in the
serialization plugin with Kotlin 1.9.20-Beta2. However, they are
unlikely to affect anyone since only generated symbols that aren't
supposed to be accessed directly were touched.